Use Helper structure to call GetGenericIssuer everywhere
