+++ pkg/controller/certificates/controller.go
@@ -39,6 +39,7 @@ import (
 
 type Controller struct {
 	*controllerpkg.Context
+	helper controllerpkg.Helper
 
 	// To allow injection for testing.
 	syncHandler func(ctx context.Context, key string) error
@@ -94,6 +95,7 @@ func New(ctx *controllerpkg.Context) *Controller {
 	ordersInformer.Informer().AddEventHandler(&controllerpkg.BlockingEventHandler{WorkFunc: ctrl.handleOwnedResource})
 	ctrl.syncedFuncs = append(ctrl.syncedFuncs, ordersInformer.Informer().HasSynced)
 
+	ctrl.helper = controllerpkg.NewHelper(ctrl.issuerLister, ctrl.clusterIssuerLister)
 	ctrl.metrics = metrics.Default
 
 	return ctrl
